ChatBotApplication/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==4.0.3
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.3.0
fonttools==4.60.1
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.7
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.79
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langsmith==0.4.37
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.9.4
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.11.0
numpy==2.0.2
openai==2.7.2
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pillow==11.3.0
propcache==0.4.1
protobuf==6.33.1
pyarrow==21.0.0
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
streamlit==1.50.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
